js logic for the apps windows
