+++ src/styles/main.css
@@ -1,5 +1,6 @@
 main .google-app {
-  display: block;
+  display: none;
+  display: none;
   position: absolute;
   top: 4rem;
   left: 5rem;
@@ -38,7 +39,7 @@ main .google-app__header--exp {
   transform: rotate(45deg);
 }
 main .music-app {
-  display: block;
+  display: none;
   position: absolute;
   top: 7rem;
   left: 20rem;
@@ -76,30 +77,30 @@ main .music-app__header--exp {
   border-radius: 50%;
   transform: rotate(45deg);
 }
-main .some-other-app {
-  display: block;
+main .photos-app {
+  display: none;
   position: absolute;
   top: 10rem;
   left: 35rem;
-  background-color: rgb(80, 216, 68);
+  background-color: rgb(40, 140, 56);
   width: 50rem;
   height: 30rem;
   border-radius: 10px;
 }
-main .some-other-app__header {
+main .photos-app__header {
   width: 100%;
   height: 2.7rem;
   border-radius: 10px 10px 0 0;
   background-color: rgb(229, 228, 228);
 }
-main .some-other-app__header--close {
+main .photos-app__header--close {
   margin: 1rem 0 0 1rem;
   width: 0.7rem;
   background-color: rgb(251, 72, 72);
   border: 0.1px solid rgb(194, 55, 55);
   border-radius: 50%;
 }
-main .some-other-app__header--min {
+main .photos-app__header--min {
   width: 0.5rem;
   height: 0.3rem;
   margin: 0 0.3rem;
@@ -108,7 +109,7 @@ main .some-other-app__header--min {
   border: 0.1px solid rgb(206, 182, 2);
   border-radius: 50%;
 }
-main .some-other-app__header--exp {
+main .photos-app__header--exp {
   width: 0.7rem;
   background-color: rgb(2, 194, 2);
   border: 0.1px solid rgb(1, 140, 1);
