+++ src/styles/styles.css
@@ -81,11 +81,8 @@ body footer .icons-area .icon--firefox {
 body footer .icons-area .icon--safari {
   width: 2.8rem;
 }
-body footer .icons-area .icon--slack {
-  width: 1.7rem;
-  background-color: white;
-  padding: 0.3rem;
-  border-radius: 10px;
+body footer .icons-area .icon--photos {
+  width: 2.4rem;
   margin: auto;
 }
 body footer .icons-area .icon--visual {
