+++ src/styles/styles.scss
@@ -92,11 +92,8 @@ body {
         &--safari {
           width: 2.8rem;
         }
-        &--slack {
-          width: 1.7rem;
-          background-color: white;
-          padding: 0.3rem;
-          border-radius: 10px;
+        &--photos {
+          width: 2.4rem;
           margin: auto;
         }
         &--visual {
